+++ packages/app/src/hooks/useOffline.ts
@@ -1,0 +1,25 @@
+import { useState, useEffect, useCallback } from 'react'
+
+const useOffline = () => {
+  const [offline, setOffline] = useState(
+    typeof navigator === 'undefined' ? false : !navigator.onLine
+  )
+
+  const handleNetworkStatusChange = useCallback(() => {
+    setOffline(!navigator.onLine)
+  }, [])
+
+  useEffect(() => {
+    window.addEventListener('online', handleNetworkStatusChange)
+    window.addEventListener('offline', handleNetworkStatusChange)
+
+    return () => {
+      window.removeEventListener('online', handleNetworkStatusChange)
+      window.removeEventListener('offline', handleNetworkStatusChange)
+    }
+  }, [handleNetworkStatusChange])
+
+  return offline
+}
+
+export default useOffline